kv: Use strict types for common fields
This PR introduces 3 new typed fields in mvcc.go:
RaftTerm, RaftIndex and LeaseSequence. These fields were previously just
unit64 throughout the code and this made the code harder to read and
risked incorrect conversions.

Epic: none

Release note: None

+++ pkg/kv/kvserver/store_raft.go
@@ -233,8 +233,8 @@ func (s *Store) uncoalesceBeats(
 			Type:   msgT,
 			From:   uint64(beat.FromReplicaID),
 			To:     uint64(beat.ToReplicaID),
-			Term:   beat.Term,
-			Commit: beat.Commit,
+			Term:   uint64(beat.Term),
+			Commit: uint64(beat.Commit),
 		}
 		beatReqs[i] = kvserverpb.RaftMessageRequest{
 			RangeID: beat.RangeID,
